Switch the config/initializer changes back closer to how they were.
I think James A & I were following a bit of a red-herring with this config stuff yesterday. It turns out the app deployed to Heroku was pointing at the Linode Prism instance which had license issues. Also the worker process is being run using the Rails runner, so the Rails environment & initializers were being loaded anyway.