Improved test coverage for transfer DSL script and fixed a small bug.
We are still only considering a scenario with two endpoints.

We noticed that the last wait_for(Connfu::Event::Hangup) was not having any effect and in fact the script was exiting immediately after the call to the #hangup method. This was because the #hangup method was setting the @finished flag. We've inlined the call to the #hangup method but removed the setting of this flag.

We've also made the incoming events more realistic by supplying iq IDs.