Waiting until Prism media is ready is already handled by the DSL.
As we understand it, we need to pause between a call being answered and then joining the call in order to give Prism a chance to get everything in order. Rather than using `sleep 1` this behaviour was encapsulated into a method called `wait_because_of_tropo_bug_133` to better reveal intention.

We don't believe this pause should be invoked explicitly in DSL scripts any more, because it is automatically invoked before any `on_answer` behaviour is invoked (see connfu/dsl.rb:133).

The explicit `sleep 1` also makes the tests run slowly; in the event that the DSL cannot automatically manage this, the `wait_because_of_tropo_bug_133` should always be used, as this is a no-op in the test environment.